requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
graphene==2.1.9
Flask-GraphQL==2.0.1
pymongo==4.6.0
mongoengine==0.27.0
Werkzeug==3.0.1 